pyproject.toml
[tool.poetry]
name = "cdk-example"
version = "0.1.0"
description = ""
authors = ["Desiree Pelletier <pelletier.desiree@gmail.com>"]
readme = "README.md"
packages = [{include = "cdk_example"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.2"
pydantic = "^1.10.6"
pandas = "^1.5.3"
XlsxWriter = "^3.0.9"


[tool.poetry.dev-dependencies]
# CDK
"aws-cdk.aws-lambda-python-alpha" = "^2.67.0a0"
aws-cdk-lib = "^2.67.0"
constructs = "^10.1.271"
# DEV
pytest = "^7.2.2"
python-dotenv = "^1.0.0"
isort = "*"
boto3 = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
